+++ app/components/chart/daily.js
@@ -39,6 +39,7 @@ const SMBGTooltip = vizComponents.SMBGTooltip;
 const CBGTooltip = vizComponents.CBGTooltip;
 const FoodTooltip = vizComponents.FoodTooltip;
 const PhysicalTooltip = vizComponents.PhysicalTooltip;
+const ReservoirTooltip = vizComponents.ReservoirTooltip;
 
 import Header from './header';
 import Footer from './footer';
@@ -70,6 +71,8 @@ const DailyChart = translate()(class DailyChart extends Component {
     onCarbOut: React.PropTypes.func.isRequired,
     onPhysicalHover: React.PropTypes.func.isRequired,
     onPhysicalOut: React.PropTypes.func.isRequired,
+    onReservoirHover: React.PropTypes.func.isRequired,
+    onReservoirOut: React.PropTypes.func.isRequired,
   };
 
   constructor(props) {
@@ -91,6 +94,8 @@ const DailyChart = translate()(class DailyChart extends Component {
       'onCarbOut',
       'onPhysicalHover',
       'onPhysicalOut',
+      'onReservoirHover',
+      'onReservoirOut',
     ];
 
     this.log = bows('Daily Chart');
@@ -319,6 +324,8 @@ class Daily extends Component {
                 onCarbOut={this.handleCarbOut}
                 onPhysicalHover={this.handlePhysicalHover}
                 onPhysicalOut={this.handlePhysicalOut}
+                onReservoirHover={this.handleReservoirHover}
+                onReservoirOut={this.handleReservoirOut}
                 ref="chart" />
             </div>
           </div>
@@ -393,6 +400,16 @@ class Daily extends Component {
           }}
           side={this.state.hoveredPhysical.side}
           physicalActivity={this.state.hoveredPhysical.data}
+          bgPrefs={this.props.bgPrefs}
+          timePrefs={this.props.timePrefs}
+        />}
+        {this.state.hoveredReservoir && <ReservoirTooltip
+          position={{
+            top: this.state.hoveredReservoir.top,
+            left: this.state.hoveredReservoir.left
+          }}
+          side={this.state.hoveredReservoir.side}
+          reservoir={this.state.hoveredReservoir.data}
           bgPrefs={this.props.bgPrefs}
           timePrefs={this.props.timePrefs}
         />}
@@ -614,6 +631,29 @@ class Daily extends Component {
   handlePhysicalOut = () => {
     this.setState({
       hoveredPhysical: false
+    });
+  };
+
+  handleReservoirHover = reservoir => {
+    var rect = reservoir.rect;
+    // range here is -12 to 12
+    var hoursOffset = sundial.dateDifference(reservoir.data.normalTime, this.state.datetimeLocation, 'h');
+    reservoir.top = rect.top + (rect.height / 2)
+    if(hoursOffset > 5) {
+      reservoir.side = 'left';
+      reservoir.left = rect.left;
+    } else {
+      reservoir.side = 'right';
+      reservoir.left = rect.left + rect.width;
+    }
+    this.setState({
+      hoveredReservoir: reservoir
+    });
+  }
+
+  handleReservoirOut = () => {
+    this.setState({
+      hoveredReservoir: false
     });
   };