+++ CHANGELOG.md
@@ -15,6 +15,9 @@ It is based on Tidepool Blip 1.27.
 ### Added
  - PT-865 Display physical activity in Bolus/Food section
 
+### Added
+ - PT-883 Display reservoir change in daily BG section
+
 ## 0.10.0 - 2019-12-09 
 ### Added
  - PT-844 Integrate Tideline 1.4.0 with new logo/picto for insulin sites